+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{Benazzo:2014,
     Author = {{Benazzo}, A. and {Panziera}, A. and {Bertorelle}, G.},
     Journal = {Ecol Evol},
-    Title = {{4P: Fast computing of population genetics statistics from large DNA polymorphism panels}}
+    Title = {{4P: Fast computing of population genetics statistics from large DNA polymorphism panels}},
     Year = 2014
 }
 
@@ -21,15 +21,15 @@
 
 @article{Gravel:2011,
     Author = {{Gravel}, S. and {Henn}, B.~M. and {Gutenkunst}, R.~N. and {Indap}, A.~R. and {Marth}, G.~T. and {Clark}, A.~G. and {Yu}, F. and {Gibbs}, R.~A. and {The 1000 Genomes Project} and {Bustamante}, C.~D.},
-    Journal = {Proc Natl Acad Sci USA}
-    Title = {{Demographic history and rare allele sharing among human populations}}
+    Journal = {Proc Natl Acad Sci USA},
+    Title = {{Demographic history and rare allele sharing among human populations}},
     Year = 2011
 }
 
 @article{Haldane:1924,
     Author = {{Haldane}, J.~B.~S.},
     Journal = {Trans Camb Phil Soc},
-    Title = {{A mathematical theory of natural and artificial selection. Part I.}}
+    Title = {{A mathematical theory of natural and artificial selection. Part I.}},
     Year = 1924
 }
 
