### 2019-07-01 82d34af
20190701 Update paper.md
