+++ paper/paper.md
@@ -39,32 +39,32 @@ Analyzing natural selection is critically important in population genetics [@Hal
 
 SeleDiff implements a probabilistic method from our previous study [@He:2015]. In this approach, we introduced logarithm odds ratios of allele frequencies to measure differences in selective pressures. For a bi-allelic locus in the population $i$, let $p_i\left(t\right)$ and $q_i\left(t\right)$ denote the derived and ancestral allele frequencies at time $t$. We define the absolute fitness of the derived and ancestral alleles as $w_D$ and $w_A$. The relative fitness becomes
 
-$$e^s = \frac{w_D}{w_A}$$,
+$$e^s = \frac{w_D}{w_A},$$
 
 where $s$ is the (genic) selection coefficient. The selection (coefficient) difference between populations $i$ and $j$ is
 
-$$d_{ij} = s_i - s_j = \frac{1}{t}\left[\ln\frac{p_i\left(t\right)/q_i\left(t\right)}{p_j\left(t\right)/q_j\left(t\right)} + \Omega\right] = \frac{1}{t}\left(\ln\text{OR}+\Omega\right)$$,
+$$d_{ij} = s_i - s_j = \frac{1}{t}\left[\ln\frac{p_i\left(t\right)/q_i\left(t\right)}{p_j\left(t\right)/q_j\left(t\right)} + \Omega\right] = \frac{1}{t}\left(\ln\text{OR}+\Omega\right),$$
 
 where $\text{OR}$ stands for odds ratio; $\Omega$ approximately follows a normal distribution with a mean of zero and reflects the uncertainty of allele frequencies caused by factors other than selection; $t$ is the divergence time from populations $i$ and $j$ to their most recent common ancestor. Thus, the expectation and variance of $d_{ij}$ are
 
 $$\begin{aligned}
 \text{E}\left(d_{ij}\right) &= \text{E}\left(s_i - s_j\right) = \frac{1}{t}\ln\text{OR} \\
 \text{var}\left(d_{ij}\right) &= \frac{1}{t^2}\left[\text{var}\left(\text{OR}\right)+\text{var}\left(\Omega\right)\right]
-\end{aligned}$$.
+\end{aligned}.$$
 
 Given a dataset with $n$ loci, we can estimate $\text{var}\left(\Omega\right)$ as
 
-$$\text{median}\left{\frac{\ln^2\text{OR}_k}{0.455}-\text{var}\left(\ln\text{OR}_k\right)\right}, 1\leq k \leq n$$,
+$$\text{median}\left\{\frac{\ln^2\text{OR}_k}{0.455}-\text{var}\left(\ln\text{OR}_k\right)\right\}, 1\leq k \leq n,$$
 
 where $\text{var}\left(\ln\text{OR}\right)\approx\frac{1}{N_i\hat{p}_i\left(t\right)}+\frac{1}{N_i\hat{q}_i\left(t\right)}+\frac{1}{N_j\hat{p}_j\left(t\right)+\frac{1}{N_j\hat{q}_j\left(t\right)}$. Here, $N_i$ and $N_j$ are the sample sizes of populations $i$ and $j$. We add 0.5 to allele counts less than 5 for continuity correction. To test the selection differences in a locus, we proposed a statistic:
 
-$$\delta = \frac{\left[\text{E}\left(d_{ij}\right)\right]^2}{\text{var}\left(d_{ij}\right)}$$,
+$$\delta = \frac{\left[\text{E}\left(d_{ij}\right)\right]^2}{\text{var}\left(d_{ij}\right)},$$
 
 where $\delta$ follows a central $\chi^2$-distribution with one degree of freedom in the absence of selection differences.
 
 ![Fig. 1](https://raw.githubusercontent.com/xin-huang/SeleDiff/master/figures/Fig1.png)
 
[…]
