### 2019-07-01 b77a4b7
20190701 Update paper.md
