### 2019-07-01 e8cc370
20190701 Update paper.md
