+++ paper/paper.md
@@ -62,21 +62,15 @@ $$\delta = \frac{\left[\text{E}\left(d_{ij}\right)\right]^2}{\text{var}\left(d_{
 
 where $\delta$ follows a central $\chi^2$-distribution with one degree of freedom in the absence of selection differences.
 
-![Fig. 1](https://raw.githubusercontent.com/xin-huang/SeleDiff/master/figures/Fig1.png)
+![Fig. 1. The demographic models in simulation. (A) The Basic Models. (B) The Bottleneck Models. In these models, we defined the bottleneck strength as $f=N_A/N_B$, where $N_A$ and $N_B$ are the population sizes after and before the bottleneck, respectively. (C) The Expansion Models. In these models, we defined the expansion strength as $a=\ln\left(\frac{N_t}{N_0}\right)/t$, where $N_0$ and $N_t$ are the population sizes before and after the expansion, respectively; $t$ is the time for expansion. (D) The Substructure Models. In these models, we assumed at least one population diverged into two subpopulations at the 1000th generation, and the beneficial allele became neutral ($s = 0$) in one of the subpopulations after divergence. (E) The Migration Models. (F) The Complex Models. Populations are abbreviated as follows: AFR, Africans; EUR, Europeans; EAS, East Asians.](https://raw.githubusercontent.com/xin-huang/SeleDiff/master/figures/Fig1.png)
 
-**Fig. 1. The demographic models in simulation.** (A) The Basic Models. (B) The Bottleneck Models. In these models, we defined the bottleneck strength as $f=N_A/N_B$, where $N_A$ and $N_B$ are the population sizes after and before the bottleneck, respectively. (C) The Expansion Models. In these models, we defined the expansion strength as $a=\ln\left(\frac{N_t}{N_0}\right)/t$, where $N_0$ and $N_t$ are the population sizes before and after the expansion, respectively; $t$ is the time for expansion. (D) The Substructure Models. In these models, we assumed at least one population diverged into two subpopulations at the 1000th generation, and the beneficial allele became neutral ($s = 0$) in one of the subpopulations after divergence. (E) The Migration Models. (F) The Complex Models. Populations are abbreviated as follows: AFR, Africans; EUR, Europeans; EAS, East Asians. 
[…]
