### 2019-07-01 100257b
20190701 Update paper.md
