### 2019-07-01 49a8972
20190701 Update paper.md
