+++ paper/paper.md
@@ -62,15 +62,15 @@ $$\delta = \frac{\left[\text{E}\left(d_{ij}\right)\right]^2}{\text{var}\left(d_{
 
 where $\delta$ follows a central $\chi^2$-distribution with one degree of freedom in the absence of selection differences.
 
-![The demographic models in simulation. (A) The Basic Models. (B) The Bottleneck Models. In these models, we defined the bottleneck strength as $f=N_A/N_B$, where $N_A$ and $N_B$ are the population sizes after and before the bottleneck, respectively. (C) The Expansion Models. In these models, we defined the expansion strength as $a=\ln\left(\frac{N_t}{N_0}\right)/t$, where $N_0$ and $N_t$ are the population sizes before and after the expansion, respectively; $t$ is the time for expansion. (D) The Substructure Models. In these models, we assumed at least one population diverged into two subpopulations at the 1000th generation, and the beneficial allele became neutral ($s = 0$) in one of the subpopulations after divergence. (E) The Migration Models. (F) The Complex Models. Populations are abbreviated as follows: AFR, Africans; EUR, Europeans; EAS, East Asians.](https://raw.githubusercontent.com/xin-huang/SeleDiff/master/figures/Fig1.png)
+![The demographic models in simulation.](https://raw.githubusercontent.com/xin-huang/SeleDiff/master/figures/Fig1.png)
 
 We evaluated SeleDiff in different demographic models (Figure 1) simulated by SLiM 2 [@Haller:2017]. In Models 1–70, we assume larger selection coefficients in Population 1 than in Population 2 (Figure 1A–E).  Without migration, SeleDiff accurately estimates selection differences ranging from 0 to 0.002/generation in scenarios with different population sizes (Figure 2A, Models 1–9). The estimated differences (Figure 2A, Models 10–17) are slightly smaller in scenarios with low initial frequencies (≤ 0.02) of the selective allele or long divergence times (≥ 5000 generations), because alleles with low initial frequencies are easily lost regardless of their selection coefficients, and alleles with small selection coefficients can reach high frequencies with long enough time.  SeleDiff is affected little by time-varied population sizes (Figure 2A, Models 18–37), except for extremely severe bottlenecks in populations under less selective pressures (Figure 2A, Model 23). In Models 38–46 (Figure 2A), populations diverge into subpopulations, and selection stops in one of these subpopulations. If we ignore their structures, then the estimated differences diminish because SeleDiff treats all the individuals in a group homogenously.  Therefore, it is important to select samples carefully and interpret results cautiously. In models with moderate migration rates (0.00001–0.0001/generation), the estimated differences are only slightly smaller than the given values, whereas strong migration reduces differences between populations (Figure 2B, Models 47–70), a well-known phenomenon in population genetics [@Crow:2009]. SeleDiff also works well in complex models (Figure 2B, Model 1a–6d) involving multiple demographic events from human evolution [@Gravel:2011]. Thus, SeleDiff is robust in various demographic models, and indicates the lower bounds of differences in selective pressures when migration or substructure exists.
 
[…]
