### 2019-07-01 72f7dcf
20190701 Update paper.md
