### 2019-07-01 8a987b3
20190701 Update paper.md
