### 2019-07-02 4aa0552
20190702 Update paper.bib
