+++ paper/paper.bib
@@ -2,7 +2,10 @@
     Author = {{Benazzo}, A. and {Panziera}, A. and {Bertorelle}, G.},
     Journal = {Ecol Evol},
     Title = {{4P: Fast computing of population genetics statistics from large DNA polymorphism panels}},
-    Year = 2014
+    Year = 2014,
+    volume = {5},
+    pages = {172–175},
+    doi = {10.1002/ece3.1261}
 }
 
 @book{Crow:2009,
@@ -14,7 +17,7 @@
 
 @book{Dunning:2014,
     Author = {{Dunning}, T. and {Friedman}, E.},
-    Booktitle = {Practical machine learning: A new look at anormaly detection},
+    Booktitle = {Practical machine learning: A new look at anomaly detection},
     Publisher = {O'Reilly Media, Inc.},
     Year = 2014
 }
@@ -23,47 +26,67 @@
     Author = {{Gravel}, S. and {Henn}, B.~M. and {Gutenkunst}, R.~N. and {Indap}, A.~R. and {Marth}, G.~T. and {Clark}, A.~G. and {Yu}, F. and {Gibbs}, R.~A. and {The 1000 Genomes Project} and {Bustamante}, C.~D.},
     Journal = {Proc Natl Acad Sci USA},
     Title = {{Demographic history and rare allele sharing among human populations}},
-    Year = 2011
+    Year = 2011,
+    volume = {108},
+    pages = {11983-11988},
+    doi = {10.1073/pnas.1019276108}
 }
 
 @article{Haldane:1924,
     Author = {{Haldane}, J.~B.~S.},
     Journal = {Trans Camb Phil Soc},
     Title = {{A mathematical theory of natural and artificial selection. Part I.}},
-    Year = 1924
+    Year = 1924,
+    volume = {23},
+    pages = {19-41},
 }
 
 @article{Haller:2017,
     Author = {{Haller}, B.~C. and {Messer}, P.~W.},
     Journal = {Mol Biol Evol},
     Title = {{SLiM 2: Flexible, interactive forward genetic simulations}},
-    Year = 2017
+    Year = 2017,
+    volume = {34},
+    pages = {230-240},
+    doi = {10.1093/molbev/msw211}
 }
 
 @article{He:2015,
     Author = {{He}, Y. and {Wang}, M. and {Huang}, X. and {Li}, R. and {Xu}, H. and {Xu}, S. and {Jin}, L.},
     Journal = {Genome Res},
     Title = {{A probabilistic method for testing and estimating selection differences between populations}},
-    Year = 2015
+    Year = 2015,
+    volume = {25},
+    pages = {1903-1909},
+    doi = {10.1101/gr.192336.115}
 }
 
 @article{Szpeich:2014,
     Author = {{Szpeich}, Z.~A. and {Hernandez}, R.~D.},
     Journal = {Mol Biol Evol},
     Title = {{selscan: An efficient multithreaded program to perform EHH-based scans for positive selection}},
-    Year = 2014
+    Year = 2014,
+    volume = {31},
+    pages = {2824-2827},
+    doi = {10.1093/molbev/msu211}
 }
 
 @article{Thurman:2016,
     Author = {{Thurman}, T.~J. and {Barrett}, R.~D.},
     Journal = {Mol Ecol},
     Title = {{The genetic consequences of selection in natural populations}},
-    Year = 2016
+    Year = 2016,
+    volume = {25},
+    pages = {1429-1448},
+    doi = {10.1111/mec.13559}
 }
 
 @article{Vitti:2013,
     Author = {{Vitti}, J.~J. and {Grossman}, S.~R. and {Sabeti}, P.~C.},
     Journal = {Annu Rev Genet},
[…]
