### 2019-07-03 #1545 @whedon
[REVIEW]: SeleDiff: A fast and scalable tool for estimating and testing selection differences between populations
**Submitting author:** @xin-huang (<a href="http://orcid.org/0000-0002-9918-9602">Xin Huang</a>)
**Repository:** <a href="https://github.com/xin-huang/SeleDiff" target ="_blank">https://github.com/xin-huang/SeleDiff</a>
**Version:** v1.1
