+++ paper/paper.bib
@@ -10,14 +10,14 @@
 
 @book{Crow:2009,
     Author = {{Crow}, J.~F. and {Kimura}, M.},
-    Booktitle = {An introduction to population genetics theory},
+    Title = {{An Introduction to Population Genetics Theory}},
     Publisher = {The Blackburn Press},
     Year = 2009
 }
 
 @book{Dunning:2014,
     Author = {{Dunning}, T. and {Friedman}, E.},
-    Booktitle = {Practical machine learning: A new look at anomaly detection},
+    Title = {{Practical machine learning: A new look at anomaly detection}},
     Publisher = {O'Reilly Media, Inc.},
     Year = 2014
 }
