+++ README.md
@@ -14,6 +14,7 @@
 - `SeleDiff` implements a probabilistic method for estimating and testing selection (coefficient) differences between populations<sup>1</sup>.
 - If you have any problem, please feel free to contact xin.huang07@gmail.com, or open an [issue](https://github.com/xin-huang/SeleDiff/issues) in this repository.
 - If you are interested in contributing to `SeleDiff`, please feel free to clone and modify it.
+- If you would like to reproduce our simulation, please check the codes in `./appendix`.
 - For more details, please see the manual in `./docs`.
 
 ## Installation
