+++ paper/paper.bib
@@ -17,7 +17,7 @@
 
 @book{Dunning:2014,
     Author = {{Dunning}, T. and {Friedman}, E.},
-    Title = {{Practical machine learning: A new look at anomaly detection}},
+    Title = {{Practical Machine Learning: A New Look at Anomaly Detection}},
     Publisher = {O'Reilly Media, Inc.},
     Year = 2014
 }
