### 2019-07-04 7de77a2
20190704 Update paper.bib
