+++ paper/paper.bib
@@ -10,6 +10,7 @@
 
 @book{Crow:2009,
     Author = {{Crow}, J.~F. and {Kimura}, M.},
+    Booktitle = {An Introduction to Population Genetics Theory},
     Title = {{An Introduction to Population Genetics Theory}},
     Publisher = {The Blackburn Press},
     Year = 2009
@@ -17,6 +18,7 @@
 
 @book{Dunning:2014,
     Author = {{Dunning}, T. and {Friedman}, E.},
+    Booktitle = {Practical Machine Learning: A New Look at Anomaly Detection},
     Title = {{Practical Machine Learning: A New Look at Anomaly Detection}},
     Publisher = {O'Reilly Media, Inc.},
     Year = 2014
