### 2019-07-08 #1545 @stulacy
Can you please add the DOI for Haldane (1924)?
