### 2019-07-08 cd8db7d
20190708 Update paper.bib
