+++ paper/paper.bib
@@ -41,6 +41,7 @@
     Year = 1924,
     volume = {23},
     pages = {19-41},
+    doi = {10.1007/BF02459574}
 }
 
 @article{Haller:2017,
