### 2019-07-08 e4b1fa4
20190708 Update paper
