+++ paper/paper.bib
@@ -34,13 +34,13 @@
     doi = {10.1073/pnas.1019276108}
 }
 
-@article{Haldane:1924,
+@article{Haldane:1990,
     Author = {{Haldane}, J.~B.~S.},
-    Journal = {Trans Camb Phil Soc},
-    Title = {{A mathematical theory of natural and artificial selection. Part I.}},
-    Year = 1924,
-    volume = {23},
-    pages = {19-41},
+    Journal = {Bltn Mathcal Biology},
+    Title = {{A mathematical theory of natural and artificial selection—I.}},
+    Year = 1990,
+    volume = {52},
+    pages = {209},
     doi = {10.1007/BF02459574}
 }
 
