### 2019-07-08 0ae27e2
20190708 Update paper.bib
