+++ paper/paper.bib
@@ -40,7 +40,7 @@
     Title = {{A mathematical theory of natural and artificial selection—I.}},
     Year = 1990,
     volume = {52},
-    pages = {209},
+    pages = {209–240},
     doi = {10.1007/BF02459574}
 }
 
