### 2019-07-10 #1545 @xin-huang
Hi @suresh-lab
I have revised the community guidelines in README:

"If you are interested in contributing to SeleDiff, please feel free to clone and modify it. You should include unit tests for your modified codes. Besides, you can edit build.gradle to include new dependencies. After your modification, please send a GitHub Pull Request with a clear list of what you've done."

Please let me know whether it is appropriate.
