### 2019-07-15 #1545 @lpantano
@xin-huang, it seems there is another DOI missing:

- https://doi.org/10.2307/1529706 may be missing for title: An Introduction to Population Genetics Theory

can you fix it?

And after that, can you make a zenodo archive and give me the DOI? you need to make sure the title and authors there match your paper.

thanks!
