### 2019-07-15 #1545 @xin-huang
Hi @lpantano

I think this https://doi.org/10.2307/1529706 is a book review for the book I cited.  The book, An Introduction to Population Genetics Theory, is a old book published in 1970s (https://www.amazon.com/Introduction-Population-Genetics-Theory/dp/1932846123), I cannot find a doi related to this book.

The DOI for the zenodo archive is 10.5281/zenodo.3323833

Thank you very much.
