### 2019-07-17 #1545 @lpantano
Hi @openjournals/joss-eics, this is accepted. The only thing is that reference that doesn't have a DOI because it is too old. I don't know if there is a specific way to deal with that. Thanks!
