### 2019-07-17 #1545 @danielskatz
Thanks - I'll do the processing tonight or tomorrow

Don't worry about the DOI, whedon provides guidance but is not always right, and everything doesn't have to have a DOI; we just want to make sure that for things that do have a DOI, the DOI is part of the reference.
