### 2019-07-17 cc22c6f
minor wording changes
