+++ paper/paper.md
@@ -29,11 +29,11 @@ bibliography: paper.bib
 
 # Summary
 
-Detecting and quantifying selection is a classical task in population genetics. Over the last two decades, many studies detected selection signals in genomes. Few studies, however, quantified differences in selective pressures  between populations due to the lack of efficient tools. Here we implemented an open-source software, SeleDiff, with an established probabilistic method to estimate and test differences in selective pressures between populations. Extensive simulation revealed that SeleDiff is robust in various demographic models, as well as fast and scalable for analyzing large-scale genomic datasets. Thus, SeleDiff is helpful for analyzing selection as genomic datasets grow, and is available at https://github.com/xin-huang/SeleDiff.
+Detecting and quantifying selection is a classical task in population genetics. Over the last two decades, many studies detected selection signals in genomes. Few studies, however, quantified differences in selective pressures between populations, due to the lack of efficient tools. Here we implemented an open-source software package, SeleDiff, with an established probabilistic method to estimate and test differences in selective pressures between populations. Extensive simulation revealed that SeleDiff is robust in various demographic models, as well as fast and scalable for analyzing large-scale genomic datasets. Thus, SeleDiff is helpful for analyzing selection as genomic datasets grow, and is available at https://github.com/xin-huang/SeleDiff.
 
 # Introduction
 
-Analyzing natural selection is critically important in population genetics [@Haldane:1990]. In the past 20 years, researchers have learned extensively about selection signals in genomic data [@Vitti:2013], but a deeper understanding of selection strength has remained elusive [@Thurman:2016].This is particularly due to difficulties in estimating selective pressures using empirical data. In addition, as the amount of genomic data has increased dramatically, researchers require more efficient software for analyzing large-scale genomic datasets. To meet these computational demands, we introduced and evaluated SeleDiff, a fast and scalable tool for quantifying differences in selective pressures between populations.
+Analyzing natural selection is critically important in population genetics [@Haldane:1990]. In the past 20 years, researchers have learned extensively about selection signals in genomic data [@Vitti:2013], but a deeper understanding of selection strength has remained elusive [@Thurman:2016]. This is particularly due to difficulties in estimating selective pressures using empirical data. In addition, as the amount of genomic data has dramatically increased, researchers require more efficient software for analyzing large-scale genomic datasets. To meet these computational demands, we introduced and evaluated SeleDiff, a fast and scalable tool for quantifying differences in selective pressures between populations.
 
 # Results
 
