### 2019-07-17 #1545 @danielskatz
@xin-huang - please merge https://github.com/xin-huang/SeleDiff/pull/6 to make a couple of small changes in wording.
