### 2019-07-18 #1545 @xin-huang
Hi @danielskatz

I have merged it into the main branch. Thank you very much.
